% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{print_palette}
\alias{print_palette}
\title{Print a single palette}
\usage{
print_palette(x, ...)
}
\arguments{
\item{x}{A color palette generated by \code{tarantino_palette()} function.}

\item{...}{Single or a vector of palette objects.}
}
\description{
Print a single palette
}
