% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{tarantino_palette}
\alias{tarantino_palette}
\title{A Quentin Tarantino palette generator}
\usage{
tarantino_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{ReservoirDogs}, \code{PulpFiction},
\code{KillBillVol1}, \code{KillBillVol2},
\code{JackieBrown}, \code{DeathProof},
\code{InglouriousBasterds}, \code{DjangoUnchained}}

\item{n}{Number of colors needed. All colors are extracted and picked from the posts on \href{https://moviesincolor.com}{Movies in Color}.
If \code{n} is omitted, all colours in that palette will be generated.}

\item{type}{Either "continuous" or "discrete".}
}
\value{
A vector of colours.
}
\description{
Generate discrete or continuous color palettes from Quentin Tarantino's movies.
}
\examples{
tarantino_palette("KillBillVol1")
tarantino_palette("KillBillVol1", 4)
tarantino_palette("KillBillVol1", 4, "continuous")
}
\keyword{colors}
