% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{tarantino_palettes}
\alias{tarantino_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 8.
}
\usage{
tarantino_palettes
}
\description{
Use \code{\link{tarantino_palette}} to construct palettes of desired length.
}
\keyword{datasets}
